again
again?

This is a reference to the Super Mario World Rom Hack, " 'the ", Or more commonly known as "Coronation Day". The reference is specifically from the start of the rom hack, where a text box pops up and says "Again?". It's funny, because I had to do several commits for a single sprite sheet.

I have not played the rom hack. I apologize if my vague description is wrong.