harvester cooption
this commit also speeds up Lazy rivers general scroll speeds, allowing for more characters to float by (yes, this also speeds up bald grus battleship, he doesn't appear to fast to us, but really think about how fast he's *really* going!)

this commit also patches an issue where playing Yolked and then playing literally any other song would cause the camera to freeze on wherever it first is focused on.
(if you were curious, it's because the freeplay stage forces the "isCameraOnForcedPos" variable in PlayState to be true, and since it was not coded to be accessed from other files, I made it that way by making it a public static variable, which.. if you weren't aware, static variables do not reset their values after the thing is unloaded, so the stage would keep that value as true, this commit adds a line to the start of playstate to set it to false.