dealions
this also changes the FPS counter to have a mostly transparent background for visibility purposes, and modifies Lazy rivers background character system so that there can be multiple on screen, but not two of the same character! Also changes the "MasterEditorMenu" to be a substate rather than just a state, allowing for quicker use of it (you could use it to skip the intro, yknow for. like. debugging reasons. you wouldnt want to skip that in a regular playthrough.. right????)

also also ALSO, this commit cleans up the files, removing uneeded files. mainly some extra psych stuff that psych doesnt need or use

I think I'm gonna code the HUD back in next, I've been putting that off.