Extras menu in a single commit? Satisfactory.
assets\shared\data\extras.json controls the data from that menu. most descriptions are temporary and missing artwork, This like, (should) pretty much fully implement the entire extras menu code. It works. I've done plenty of testing.

this commit also corrects incorrect instructions on how to use the save file zip
(I dont even think that save data works, I added an option to the debug substate (On the main menu press 7!!) and scroll down to see the option that will force scores to the full week)

if anyones seeing this commit and wants to take a shot at fixing the video support, feel free, I only removed like 2 lines (and they were if statements in the input part of the code) so everything else should still be there. if you fix it make a pull request and i'd be happy